Segment add / remove support
