+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/dedup/PartitionDedupMetadataManager.java
@@ -19,14 +19,20 @@
 package org.apache.pinot.segment.local.dedup;
 
 import com.google.common.annotations.VisibleForTesting;
+import java.util.HashMap;
 import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import org.apache.pinot.common.metrics.ServerMetrics;
+import org.apache.pinot.segment.local.segment.readers.PinotSegmentColumnReader;
 import org.apache.pinot.segment.local.upsert.PartitionUpsertMetadataManager;
 import org.apache.pinot.segment.local.utils.HashUtils;
 import org.apache.pinot.segment.local.utils.RecordInfo;
 import org.apache.pinot.segment.spi.IndexSegment;
 import org.apache.pinot.spi.config.table.HashFunction;
+import org.apache.pinot.spi.data.readers.PrimaryKey;
+import org.apache.pinot.spi.utils.ByteArray;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -35,6 +41,7 @@ public class PartitionDedupMetadataManager {
   private static final Logger LOGGER = LoggerFactory.getLogger(PartitionUpsertMetadataManager.class);
 
   private final String _tableNameWithType;
+  private final List<String> _primaryKeyColumns;
   private final int _partitionId;
   private final ServerMetrics _serverMetrics;
   private final HashFunction _hashFunction;
@@ -43,16 +50,18 @@ public class PartitionDedupMetadataManager {
   @VisibleForTesting
   final ConcurrentHashMap<Object, Boolean> _primaryKeySet = new ConcurrentHashMap<>();
 
-  public PartitionDedupMetadataManager(String tableNameWithType, int partitionId, ServerMetrics serverMetrics,
-      HashFunction hashFunction) {
+  public PartitionDedupMetadataManager(String tableNameWithType, List<String> primaryKeyColumns,
+      int partitionId, ServerMetrics serverMetrics, HashFunction hashFunction) {
     _tableNameWithType = tableNameWithType;
+    _primaryKeyColumns = primaryKeyColumns;
     _partitionId = partitionId;
     _serverMetrics = serverMetrics;
     _hashFunction = hashFunction;
   }
 
-  public void addSegment(IndexSegment segment, Iterator<RecordInfo> recordInfoIterator) {
+  public void addSegment(IndexSegment segment) {
     // Add all PKs to _primaryKeySet
+    Iterator<RecordInfo> recordInfoIterator = getRecordInfoIterator(segment);
     while (recordInfoIterator.hasNext()) {
       RecordInfo recordInfo = recordInfoIterator.next();
       _primaryKeySet.put(HashUtils.hashPrimaryKey(recordInfo.getPrimaryKey(), _hashFunction), true);
@@ -61,7 +70,42 @@ public class PartitionDedupMetadataManager {
 
   public void removeSegment(IndexSegment segment) {
     // Remove all PKs from _primaryKeySet
-    // how??
+    Iterator<RecordInfo> recordInfoIterator = getRecordInfoIterator(segment);
+    while (recordInfoIterator.hasNext()) {
+      RecordInfo recordInfo = recordInfoIterator.next();
+      _primaryKeySet.remove(HashUtils.hashPrimaryKey(recordInfo.getPrimaryKey(), _hashFunction));
+    }
+  }
+
+  private Iterator<RecordInfo> getRecordInfoIterator(IndexSegment segment) {
+    Map<String, PinotSegmentColumnReader> columnToReaderMap = new HashMap<>();
+    for (String primaryKeyColumn : _primaryKeyColumns) {
+      columnToReaderMap.put(primaryKeyColumn, new PinotSegmentColumnReader(segment, primaryKeyColumn));
+    }
+    int numTotalDocs = segment.getSegmentMetadata().getTotalDocs();
+    int numPrimaryKeyColumns = _primaryKeyColumns.size();
+    return new Iterator<RecordInfo>() {
+      private int _docId = 0;
+
+      @Override
+      public boolean hasNext() {
+        return _docId < numTotalDocs;
+      }
+
+      @Override
+      public RecordInfo next() {
+        Object[] values = new Object[numPrimaryKeyColumns];
+        for (int i = 0; i < numPrimaryKeyColumns; i++) {
+          Object value = columnToReaderMap.get(_primaryKeyColumns.get(i)).getValue(_docId);
+          if (value instanceof byte[]) {
+            value = new ByteArray((byte[]) value);
+          }
+          values[i] = value;
+        }
+        PrimaryKey primaryKey = new PrimaryKey(values);
+        return new RecordInfo(primaryKey, _docId++, null);
+      }
+    };
   }
 
   public boolean checkRecordPresentOrUpdate(RecordInfo recordInfo) {
