+++ pinot-segment-local/src/main/java/org/apache/pinot/segment/local/dedup/TableDedupMetadataManager.java
@@ -18,6 +18,7 @@
  */
 package org.apache.pinot.segment.local.dedup;
 
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import org.apache.pinot.common.metrics.ServerMetrics;
@@ -27,18 +28,21 @@ import org.apache.pinot.spi.config.table.HashFunction;
 public class TableDedupMetadataManager {
   private final Map<Integer, PartitionDedupMetadataManager> _partitionMetadataManagerMap = new ConcurrentHashMap<>();
   private final String _tableNameWithType;
+  private final List<String> _primaryKeyColumns;
   private final ServerMetrics _serverMetrics;
   private final HashFunction _hashFunction;
 
-  public TableDedupMetadataManager(String tableNameWithType, ServerMetrics serverMetrics, HashFunction hashFunction) {
+  public TableDedupMetadataManager(String tableNameWithType, List<String> primaryKeyColumns,
+      ServerMetrics serverMetrics, HashFunction hashFunction) {
     _tableNameWithType = tableNameWithType;
+    _primaryKeyColumns = primaryKeyColumns;
     _serverMetrics = serverMetrics;
     _hashFunction = hashFunction;
   }
 
   public PartitionDedupMetadataManager getOrCreatePartitionManager(int partitionId) {
     return _partitionMetadataManagerMap.computeIfAbsent(partitionId,
-        k -> new PartitionDedupMetadataManager(_tableNameWithType, k, _serverMetrics,
-            _hashFunction));
+        k -> new PartitionDedupMetadataManager(_tableNameWithType, _primaryKeyColumns, k,
+            _serverMetrics, _hashFunction));
   }
 }
